testmvrnormArma {IndianVectorModelling}	R Documentation

Multivariate Normal Sampler

Description

Random number generation for a multivariate normal (Gaussian) distribution. Uses a Cholesky decomposition to propose new values from a multivariate normal distribution with a given mean vector (mu) and a specified covariance matrix (sigma).

Usage

testmvrnormArma(mu, sigma)

Arguments

mu
Mean of the multivariate normal distribution (has to be specified as a matrix with 1 row and number of columns equal to the number of variables in the multivariate distribution e.g. 2 for bivariate)

sigma
A covariance matrix associated with mu

Details

Note: Currently (for my own modelling purposes) won't return values < 0.

[Package IndianVectorModelling version 0.1.0 Index]